Feature bootstrap

+++ server/public/client.js
@@ -13,8 +13,6 @@ function andGo() {
 }
 
 function completeTaskBtn() {
-  console.log($(this).data('id'), $(this).data('bool'));
-
   const taskId = $(this).data('id');
   const isComplete = $(this).data('bool');
 
@@ -26,7 +24,16 @@ function completeTaskBtn() {
     },
   })
     .then((res) => {
-      console.log(res);
+      swal({
+        title: `Marked as ${isComplete === true ? 'Incomplete' : 'Complete'}`,
+        text: `${
+          isComplete === true
+            ? 'You Can finish this task Again!'
+            : 'You Finished the task!'
+        }`,
+        icon: 'success',
+        button: `Continue`,
+      });
       $('.table-of-toDo').empty();
       getData();
     })
@@ -36,17 +43,33 @@ function completeTaskBtn() {
 function deleteBtn() {
   const theTask = $(this).data('id');
 
-  $.ajax({
-    method: 'DELETE',
-    url: `/toDoItem/${theTask}`,
-  })
-    .then((res) => {
-      // console.log(res);
-      $('.table-of-toDo').empty();
+  swal({
+    title: 'Are you sure?',
+    text: 'Once deleted, you will not be able to recover this task!',
+    icon: 'warning',
+    buttons: true,
+    dangerMode: true,
+  }).then((willDelete) => {
+    if (willDelete) {
+      $.ajax({
+        method: 'DELETE',
+        url: `/toDoItem/${theTask}`,
+      })
+        .then((res) => {
+          // console.log(res);
+          $('.table-of-toDo').empty();
 
-      getData();
-    })
-    .catch((err) => console.error(err));
+          getData();
+        })
+        .catch((err) => console.error(err));
+
+      swal('Your task has been deleted!', {
+        icon: 'success',
+      });
+    } else {
+      swal('Your task is safe!');
+    }
+  });
 }
 
 function addTask() {
@@ -72,21 +95,36 @@ function addTask() {
 
 function sendTaskToDB(theNewTask) {
   // console.log('in send task', theNewTask);
-  $.ajax({
-    method: 'POST',
-    url: '/toDoItem/postOnServer',
-    data: {
-      newTask: theNewTask,
-    },
-  })
-    .then((res) => {
-      console.log('response from server', res);
-      // clear existing list
-      $('.table-of-toDo').empty();
-      // get data, now with added task
-      getData();
-    })
-    .catch((err) => console.error(err));
+
+  swal({
+    title: 'Are you sure?',
+    text: 'This Task will be added to your list.',
+    icon: 'info',
+    buttons: true,
+    dangerMode: true,
+  }).then((addTheTask) => {
+    if (addTheTask) {
+      $.ajax({
+        method: 'POST',
+        url: '/toDoItem/postOnServer',
+        data: {
+          newTask: theNewTask,
+        },
+      })
+        .then((res) => {
+          // clear existing list
+          $('.table-of-toDo').empty();
+          // get data, now with added task
+          getData();
+        })
+        .catch((err) => console.error(err));
+      swal('You added a Task', {
+        icon: 'success',
+      });
+    } else {
+      swal('No Task was added');
+    }
+  });
 }
 
 function clearInputs() {
@@ -115,16 +153,20 @@ function renderData(itemsList) {
   itemsList.forEach((obj) => {
     // console.log(obj);
     $('.table-of-toDo').append(`
-    <tr>
+    <tr ${
+      obj.complete === false ? `class="table-danger"` : `class="table-success"`
+    }>
         <td>${obj.task_name}</td>
-        <td>${obj.completion_time}</td>
-        <td>${obj.complete}</td>
+        <td>${obj.completion_time} hours</td>
+        <td>${obj.complete === false ? 'No' : 'Yes'}</td>
         ${
           obj.complete === false
-            ? `<td><button class="complete-task-btn" data-id="${obj.id}" data-bool="${obj.complete}">Mark Complete</button></td>`
-            : `<td><button class="complete-task-btn" data-id="${obj.id}" data-bool="${obj.complete}">Mark Incomplete</button></td>`
+            ? `<td><button class="complete-task-btn btn btn-success" data-id="${obj.id}" data-bool="${obj.complete}">Mark Complete</button></td>`
+            : `<td><button class="complete-task-btn btn btn-secondary" data-id="${obj.id}" data-bool="${obj.complete}">Mark Incomplete</button></td>`
         }
-        <td><button class="delete-btn" data-id="${obj.id}">Delete</button></td> 
+        <td><button class="delete-btn btn btn-danger" data-id="${
+          obj.id
+        }">Delete</button></td> 
       </tr>
     `);
   });
